fix: bottom banner height
