+++ app/src/main/res/values/dimens.xml
@@ -2,6 +2,7 @@
 <resources>
     <dimen name="icon_size">25dp</dimen>
     <dimen name="default_padding">5dp</dimen>
+    <dimen name="ad_banner_padding">2dp</dimen>
     <dimen name="screen_padding">10dp</dimen>
     <dimen name="screen_padding_new">15dp</dimen>
     <dimen name="ads_banner_height">60dp</dimen>
